+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### `jsonschema-generator`
+#### Fixed
+- prevent mixing of `type: null` with `const`/`enum` in order to avoid validation error when `const`/`enum` does not include `null`
+
+#### Changed
+- default `ObjectMapper` when none is given in `SchemaGeneratorConfigBuilder` constructor now enables `JsonWriteFeature.WRITE_NUMBERS_AS_STRINGS`
+
 ### `jsonschema-module-jackson`
 #### Added
 - Consider `@JsonProperty.value` override for methods